Fix remove silence bug when using remove_type of "all"
This code change fixes a bug that made remove silence create an audio clip longer than the original clip.  The problem was that for each individual clip created by split, 2*padding was added which is fine when those clips remain independent as is the case with "split", but with "all" the clips are combined and the overlap causes up to a padding length delay to be added for each split.